+++ application/modules/finance/controllers/libraries/Income.php
@@ -71,9 +71,16 @@ class Income extends MY_Controller {
 
 	public function delete()
 	{
+		$this->load->model('Check_exists_model');
+		$arrget = $this->input->get();
 		$id = $this->uri->segment(5);
-		$this->Income_Model->delete($id);
-		$this->session->set_flashdata('strSuccessMsg','Income successfully deleted.');
+		if($this->Check_exists_model->check_income($arrget['code']) > 0):
+			$this->session->set_flashdata('strMsg','Income is unable to delete. Contact administrator.');
+		else:
+			$this->Income_Model->delete($id);
+			$this->session->set_flashdata('strSuccessMsg','Income successfully deleted.');
+		endif;
+		
 		redirect('finance/libraries/income');
 	}
